low-tech fix of key and cert propagation for 2 way-ssl
key and cert as passed as file path in EMG configuration.

These files are read and saved in httpsOptions as Buffer (that is
important).

Some debug log I've added shows a buffer object as (slightly edited for clarity):

  key: <Buffer 2d 2d 2d 2d 2d 42 45 47 49 4e  ... >,

Unfortunately, the line

   opts = _.merge(opts, httpsOptions);

 in config-proxy-middleware.js copy the buffer content in an Array. The
 logs then show (slightly edited for clarity):

   key:
   [ 45,
     45,
     45,
     45,
     45,
     ... 1604 more items ],

The issue comes from lodash library.

The version required by package.json is ~3.9.3. This version is too
old. Buffer are properly supported by lodash in version 4 and above.

I've tested that _.merge works as expected with lodash 4.15.0

So to fix this issue, you can:
- update lodash version in package.json
- apply this PR which avoids using lodash

All the best