Avoid overriding headers set in onresponse
Currently, overriding a header from the target response in the source
response requires doing it in both ondata_response (when receiving the first
chunk of data) and in onend_response (for when the response doesn't contain
any data), after verifying that res.headersSent is false. That's clumsy at
best.

By filtering the copied headers with those already present in the source
response, it's possible to set them in onresponse once and for all.