Adding code for keepAliveTimeout
If requests are coming from Google Load Balancer occasionally you will see 502 response codes. This happens because the GCLB has a timeout of 10 mins. This can not be changed. The microgateway has a timeout less than what Google's timeout is so it is closing connections before the GCLB resulting in the 502s. Another change will be proposed to allow this to be configured from the configuration file.