139064652 add debug and trace levels
Add debug and trace levels to the logger for both system and event logs.
The trace can be used to print the error stacks of event logs and debug can be used to print some varibale values.
The order of logging levels is as below:
trace
debug
info
warn
error